+++ Assets/Scripts/Services/Contract/IGameSessionManager.cs
@@ -6,10 +6,10 @@ namespace CaromBilliards3D.Services
     public interface IGameSessionManager : IBaseService
     {
         public GameSessionData gameSessionData { get; set; }
-        public float GetTimePlayed();
+        public int GetTimePlayed();
         public int GetScore();
         public int GetShotsTaken();
-        public void SetTimePlayed(float timePlayed);
+        public void SetTimePlayed(int timePlayed);
         public void SetScore(int score);
         public void SetShotsTaken(int shotsTaken);
         public void LoadGameSessionData(string directoryPath, string fileName, string fileExtension = ".dat");
